Create StoreTseCaseHelpers
